requirements.txt
acme==0.28.0
asn1crypto==0.22.0
backports.csv==1.0.7
backports.shutil-get-terminal-size==1.0.0
certbot==0.28.0
certbot-apache==0.28.0
certifi==2019.6.16
chardet==3.0.4
ConfigArgParse==0.11.0
configobj==5.0.6
configparser==3.7.4
cryptography==1.9
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11
django-autocomplete-light==3.3.0rc1
django-bootstrap-datepicker-plus==3.0.5
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-import-export==1.1.0
djangorestframework==3.9.0
enum34==1.1.6
et-xmlfile==1.0.1
funcsigs==0.4
future==0.15.2
idna==2.8
ipython==5.8.0
ipython-genutils==0.2.0
jdcal==1.4.1
josepy==1.1.0
mock==1.3.0
mysqlclient==1.4.2.post1
ndg-httpsclient==0.4.2
odfpy==1.4.0
openpyxl==2.6.2
parsedatetime==2.4
pathlib2==2.3.3
pbr==1.8.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==1.0.16
ptyprocess==0.6.0
pyasn1==0.1.9
pycurl==7.43.0
Pygments==2.4.2
pygobject==3.20.0
PyICU==1.9.2
pyOpenSSL==17.3.0
pyRFC3339==1.0
python-apt==1.1.0b1+ubuntu0.16.4.5
python-augeas==0.5.0
pytz==2018.7
PyYAML==5.1.1
requests==2.22.0
requests-toolbelt==0.8.0
scandir==1.10.0
selenium==3.141.0
simplegeneric==0.8.1
six==1.12.0
tablib==0.13.0
traitlets==4.3.2
unattended-upgrades==0.1
urllib3==1.25.3
virtualenv==16.6.0
wcwidth==0.1.7
xlrd==1.2.0
xlwt==1.3.0
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
